// src/types/solution.ts
export interface Solution {
  slug: string
  title: string
  shortDescription: string
  fullDescription: string
  image: string
  iconName: string // Nome do ícone da Lucide (Smartphone, UserCheck, etc.)
  problem: string
  impact: string
  benefits: string[]
  howItWorks: string[]
  ctaText: string
}